+++ commands/critfbl.js
@@ -2,11 +2,10 @@ module.exports = {
 	name: 'critfbl',
 	aliases: ['critf'],
 	category: 'fbl',
-	description: 'Rolls for a random critical injury.'
-		+ '\nType `help crit` for more details.',
+	description: 'ccritfbl-description',
 	guildOnly: false,
 	args: false,
-	usage: '[table] [numeric]',
+	usage: '[table] [numeric|-lucky [rank]] [-private|-p] [-lang language_code]',
 	async run(args, ctx) {
 		args.unshift('fbl');
 		await ctx.bot.commands.get('crit').run(args, ctx);